feat(ML): add feature to detect gestures
The code was copied over from another project, with some changes to
allow it to be used with the i2cdev library instead of the MPU6050
library of esp.

-> The model is used to detect only 3 motions, however they are not
correctly inferred because of the way the i2cdev handles mpu6050 data.
(might require changes to make it identical to other library)