WIP(model): add model to detect if device is on barbell
The model doesnt perform well at all.

Suggested improvements:
-> get more a lot more data
-> change model architecture, LSTM doesnt work for some reason
-> use some other features. So far i have tried using FFT on the gy data
but it still seems to perform badly, however it maybe still be better
than using straight up accel and gyro values (which seem to overfit the
model a lot)

Next steps:
-> it may not be worth exploring this, as it may be hard to detect if
the device is attached to the barbell in the first place, other methods
like a mechanical switch of some sort would be much more reliable.